+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased ([changes](https://github.com/colszowka/simplecov/compare/v0.7.0...ma
 -------------------
 
   * [FEATURE] Adds support for Rails 4 command guessing.
+  * Rename adapters to "profiles" given that they are bundles of settings
 
 v0.7.1, 2012-10-12 ([changes](https://github.com/colszowka/simplecov/compare/v0.7.0...v0.7.1))
 -------------------